requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
urllib3==1.26.6
Werkzeug==2.3.6
zipp==3.16.2
